basic implementation
this is a basic implementation of a web service that can help on-board
OIDC users for kubernetes. configuring kubernetes for OIDC flows is
non-trivial.

There are some implementations coming to kubectl, but these require that
every user have the oauth2 client secret. In a perfect world, this
remains...well, secret.  This front end is currently geared towards
Auth0 integration, but should be compatible with any OIDC provider.